+++ lib/solidpod.dart
@@ -45,7 +45,7 @@ export 'src/solid/grant_permission_ui.dart' show GrantPermissionUi;
 
 export 'src/solid/common_func.dart';
 export 'src/solid/utils/app_info.dart' show AppInfo;
-export 'src/solid/utils/key_management.dart' show KeyManager;
+export 'src/solid/utils/key_helper.dart' show KeyManager;
 export 'src/solid/utils/misc.dart';
 export 'src/widgets/change_key_dialog.dart' show changeKeyPopup;
 export 'src/solid/read_pod.dart' show readPod;
