+++ lib/src/solid/utils/key_helper.dart
