+++ lib/src/widgets/change_key_dialog.dart
@@ -32,7 +32,7 @@ import 'package:flutter/material.dart';
 
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 
-import 'package:solidpod/src/solid/utils/key_management.dart';
+import 'package:solidpod/src/solid/utils/key_helper.dart';
 import 'package:solidpod/src/widgets/secret_input_form.dart';
 
 /// Displays a dialog for changing the key
